Melhora na visualização por grid, troca do score por quantidade de estrelas e outros ajustes.

+++ lib/core/domain/entities/search_item.dart
@@ -3,7 +3,7 @@ class SearchItem {
   final String id;
   final String name;
   final String description;
-  final double score;
+  final int score;
 
   SearchItem({
     required this.id,
